[Complains] Set the fucking background color of Unlock Button
This modification doesn't have any effects on GtkButton for some unknown
reasons. I don't know what happened because it's really weird. This fucking
commit is just a fucking note.

GTK3 is a bunch of shit and so is GNOME Environment.

Life is short; Stay away from GTK -- a fucking GUI programming toolkit.

Qt is the future, so, GTK3, fuck you! I don't ever want to see that shit again.